+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - Add GOV.UK frontend package and layout
 - Reduce cookie size to avoid ActionDispatch::Cookies::CookieOverflow error
 - Enable Rollbar and capture unknown errors from Auth0 using Rollbar
+- Create Providers model and seed database with known providers
 
 [unreleased]: TODO
 [keep a changelog 1.0.0]: https://keepachangelog.com/en/1.0.0/